requirements.txt
selenium<=4.13.0
pytest
python-dotenv